Track deletion UI: code cleanup.
